+++ pyproject.toml
@@ -23,7 +23,6 @@ dependencies = [
   'tabulate == 0.9.0',
   'python-dotenv >= 0.21.0',
   'langchain >= 0.0.335',
-  'agent-protocol>=1.0.1',
   'llama-index >= 0.8.49',
   'rank-bm25 >= 0.2.2',
   'tree_sitter_languages >= 1.8.0'
